+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## version 3.0.0 (Released 2021-11-07)
+## Version 4.0.0 (Released 2023-??-??)
+
+- The authorizer's runtime has been updated to Node 18.
+- The authorizer now supports multiple audience–issuer pairs.
+
+## Version 3.0.0 (Released 2021-11-07)
 
 - The authorizer's deployment size has been significantly reduced due to new depdendencies.
 - The `sub` of the JWT is supplied as the `usageIdentifierKey` for rate-limiting and other purposes.